+++ crates/tauri-runtime-wry/CHANGELOG.md
@@ -1,5 +1,37 @@
 # Changelog
 
+## \[2.11.4]
+
+### Bug Fixes
+
+- [`ba865f2ef`](https://www.github.com/tauri-apps/tauri/commit/ba865f2ef5a4c48f14ed00619699e27a90db47c5) ([#15614](https://www.github.com/tauri-apps/tauri/pull/15614) by [@Legend-Master](https://www.github.com/tauri-apps/tauri/../../Legend-Master)) Fix getting the DPI internally leaks `HDC` handles on Windows. This also improves the resizing speed on undecorated windows
+
+## \[2.11.3]
+
+### Bug Fixes
+
+- [`f1fe2d683`](https://www.github.com/tauri-apps/tauri/commit/f1fe2d683c0db87707e4ed788e5aae5f6022d913) ([#15504](https://www.github.com/tauri-apps/tauri/pull/15504) by [@Legend-Master](https://www.github.com/tauri-apps/tauri/../../Legend-Master)) Fix `cookies_for_url` can cause deadlocks if called with another windowing/webview method
+- [`056069ad0`](https://www.github.com/tauri-apps/tauri/commit/056069ad0f5b537de37ccadac08c1a89507f23f1) ([#15502](https://www.github.com/tauri-apps/tauri/pull/15502) by [@bclarke123](https://www.github.com/tauri-apps/tauri/../../bclarke123)) Fix a `RefCell` `BorrowMutError` panic on mobile: the `Resumed`/`Suspended` event branch held a `windows` borrow across the window-event handlers and the `RunEvent` callback, so any of them that created or closed a window (e.g. from a resume/suspend handler) panicked.
+
+### Dependencies
+
+- Upgraded to `tauri-utils@2.9.3`
+- Upgraded to `tauri-runtime@2.11.3`
+
+## \[2.11.2]
+
+### Dependencies
+
+- Upgraded to `tauri-utils@2.9.2`
+- Upgraded to `tauri-runtime@2.11.2`
+
+## \[2.11.1]
+
+### Dependencies
+
+- Upgraded to `tauri-utils@2.9.1`
+- Upgraded to `tauri-runtime@2.11.1`
+
 ## \[2.11.0]
 
 ### New Features